Create and utilize crypt package.
