+++ db/tacitDB.go
@@ -2,7 +2,6 @@ package db
 
 import (
 	"github.com/jinzhu/gorm"
-	"golang.org/x/crypto/bcrypt"
 )
 
 type DbUser struct {
@@ -26,17 +25,13 @@ type TacitDB interface {
 	RecordNotFound() bool
 }
 
-type TacitCrypt interface {
-	GenerateFromPassword(password []byte, cost int) ([]byte, error)
-	CompareHashAndPassword(hashedPassword, password []byte) error
-}
+
 
 type RealTacitDB struct {
 	GormDB *gorm.DB
 }
 
-type RealTacitCrypt struct {
-}
+
 
 func (db *RealTacitDB) AutoMigrate(values ...interface{}) {
 	db.GormDB.AutoMigrate(values)
@@ -72,12 +67,3 @@ func RunMigration(db TacitDB) {
 	db.AutoMigrate(&DbUser{})
 }
 
-func (crypt *RealTacitCrypt) GenerateFromPassword(password []byte, cost int) ([]byte, error) {
-	hash, err := bcrypt.GenerateFromPassword(password, cost)
-	return hash, err
-}
-
-func (crypt *RealTacitCrypt) CompareHashAndPassword(hashedPassword, password []byte) error {
-	err := bcrypt.CompareHashAndPassword(hashedPassword, password)
-	return err
-}
