+++ db/tacitDB_Mock.go
@@ -20,9 +20,6 @@ type TacitDBMock struct {
 	NoRecordFound     bool
 }
 
-type TacitCryptMock struct {
-	HasGeneratePasswordError bool
-}
 
 func (db *TacitDBMock) AutoMigrate(values ...interface{}) {
 	db.TimesAutoMigrateWasCalled++
@@ -66,18 +63,4 @@ func (db *TacitDBMock) Error() error {
 //Mocks gorm NoRecordFound
 func (db *TacitDBMock) RecordNotFound() bool {
 	return db.NoRecordFound
-}
-
-func (crypt *TacitCryptMock) GenerateFromPassword(password []byte, cost int) ([]byte, error) {
-	hash, err := bcrypt.GenerateFromPassword(password, cost)
-	if crypt.HasGeneratePasswordError {
-		return hash, fmt.Errorf("___GENERIC_CRYPT_ERROR___")
-	} else {
-		return hash, err
-	}
-}
-
-func (crypt *TacitCryptMock) CompareHashAndPassword(hashedPassword, password []byte) error {
-	err := bcrypt.CompareHashAndPassword(hashedPassword, password)
-	return err
 }
